refactor: split awake.py into bridge_state + command_handlers modules
Extract shared config (bridge_state.py) and /command handlers
(command_handlers.py) from awake.py to reduce its size from 907 to 538
lines. Uses callback injection pattern for handle_chat/_run_in_worker
to avoid circular imports. All 2338 tests pass.

Rebased version of PR #84 (session 93), updated for current codebase
including skills system, CLI provider, and colored logs changes.
