+++ koan/app/command_handlers.py
@@ -1,0 +1,396 @@
+"""Telegram bridge command handlers.
+
+All /command handlers live here. Extracted from awake.py to keep
+the main module focused on polling, chat, and outbox.
+
+This module uses callback injection for handle_chat and _run_in_worker
+to avoid circular imports with awake.py.
+"""
+
+import time
+from typing import Callable, Optional
+
+from app.bridge_log import log
+from app.bridge_state import (
+    KOAN_ROOT,
+    INSTANCE_DIR,
+    MISSIONS_FILE,
+    _get_registry,
+)
+from app.notify import send_telegram
+from app.skills import Skill, SkillContext, execute_skill
+from app.utils import (
+    parse_project as _parse_project,
+    detect_project_from_text,
+    insert_pending_mission,
+)
+
+# Callbacks injected by awake.py at startup to avoid circular imports
+_handle_chat_cb: Optional[Callable] = None
+_run_in_worker_cb: Optional[Callable] = None
+
+
+def set_callbacks(
+    handle_chat: Callable,
+    run_in_worker: Callable,
+):
+    """Inject callbacks from awake.py (called once at import time)."""
+    global _handle_chat_cb, _run_in_worker_cb
+    _handle_chat_cb = handle_chat
+    _run_in_worker_cb = run_in_worker
+
+
+# Core commands that remain hardcoded (safety-critical or bootstrap)
+CORE_COMMANDS = frozenset({
+    "help", "stop", "sleep", "resume", "skill",
+    "pause", "work", "awake", "start", "restart",  # aliases for sleep/resume
+})
+
+
+def handle_command(text: str):
+    """Handle /commands — core commands hardcoded, rest via skills."""
+    cmd = text.strip().lower()
+
+    # --- Core hardcoded commands (safety-critical / bootstrap) ---
+
+    if cmd == "/stop":
+        (KOAN_ROOT / ".koan-stop").write_text("STOP")
+        send_telegram("⏹️ Stop requested. Current mission will complete, then Kōan will stop.")
+        return
+
+    if cmd in ("/pause", "/sleep"):
+        pause_file = KOAN_ROOT / ".koan-pause"
+        if pause_file.exists():
+            send_telegram("⏸️ Already paused. /resume to unpause.")
+        else:
+            pause_file.write_text("PAUSE")
+            send_telegram("⏸️ Paused. No missions will run. /resume to unpause.")
+        return
+
+    if cmd in ("/resume", "/work", "/awake", "/start", "/restart"):
+        handle_resume()
+        return
+
+    if cmd == "/help" or cmd.startswith("/help "):
+        help_args = text.strip()[5:].strip()  # everything after "/help"
+        if help_args:
+            _handle_help_command(help_args)
+        else:
+            _handle_help()
+        return
+
+    if cmd.startswith("/skill"):
+        _handle_skill_command(text[6:].strip())
+        return
+
+    # --- Skill-based dispatch ---
+
+    # Extract command name and args from /command_name args
+    parts = text.strip().split(None, 1)
+    command_name = parts[0][1:].lower()  # strip the /
+    command_args = parts[1] if len(parts) > 1 else ""
+
+    # Aliases are handled by the skill registry (SKILL.md aliases: field)
+    # No hardcoded alias remapping needed here.
+
+    registry = _get_registry()
+    skill = registry.find_by_command(command_name)
+
+    if skill is not None:
+        _dispatch_skill(skill, command_name, command_args)
+        return
+
+    # Scoped command dispatch: /<scope>.<name> [args]
+    # e.g., /anantys.review or /core.status.ping
+    if "." in command_name:
+        resolved = registry.resolve_scoped_command(
+            command_name + (" " + command_args if command_args else "")
+        )
+        if resolved:
+            skill, subcommand, skill_args = resolved
+            _dispatch_skill(skill, subcommand, skill_args)
+            return
+
+    # Unknown command — pass to Claude as chat
+    if _handle_chat_cb:
+        _handle_chat_cb(text)
+
+
+def _dispatch_skill(skill: Skill, command_name: str, command_args: str):
+    """Dispatch a skill execution — handles worker threads and standard calls."""
+    ctx = SkillContext(
+        koan_root=KOAN_ROOT,
+        instance_dir=INSTANCE_DIR,
+        command_name=command_name,
+        args=command_args,
+        send_message=send_telegram,
+        handle_chat=_handle_chat_cb,
+    )
+
+    # Worker thread for blocking skills (calls Claude or external services)
+    if skill.worker:
+        def _run_skill():
+            result = execute_skill(skill, ctx)
+            if result:
+                send_telegram(result)
+        if _run_in_worker_cb:
+            _run_in_worker_cb(_run_skill)
+        return
+
+    # Standard skill execution
+    result = execute_skill(skill, ctx)
+    if result is not None:
+        send_telegram(result)
+
+
+def _handle_skill_command(args: str):
+    """Handle /skill — list skills or invoke a specific one.
+
+    Usage:
+        /skill                    — list all skills
+        /skill core               — list skills in scope 'core'
+        /skill core.status        — invoke core/status skill
+        /skill core.status.ping   — invoke subcommand 'ping' of core/status
+    """
+    registry = _get_registry()
+
+    if not args:
+        # List non-core skills grouped by scope (core skills are in /help)
+        non_core = [s for s in registry.list_all() if s.scope != "core"]
+        if not non_core:
+            send_telegram("ℹ️ No extra skills loaded. Core skills are listed in /help.")
+            return
+
+        parts = ["Available Skills\n"]
+        non_core_scopes = sorted(set(s.scope for s in non_core))
+        for scope in non_core_scopes:
+            parts.append(f"{scope}")
+            for skill in registry.list_by_scope(scope):
+                for cmd in skill.commands:
+                    desc = cmd.description or skill.description
+                    parts.append(f"  /{scope}.{cmd.name} -- {desc}")
+            parts.append("")
+
+        parts.append("Use: /<scope>.<name> [args]")
+        parts.append("Core skills are listed in /help.")
+        send_telegram("\n".join(parts))
+        return
+
+    # Parse skill reference: scope.name[.subcommand] [args]
+    ref_parts = args.split(None, 1)
+    ref = ref_parts[0]
+    skill_args = ref_parts[1] if len(ref_parts) > 1 else ""
+
+    segments = ref.split(".")
+
+    if len(segments) == 1:
+        # Just a scope — list skills in that scope
+        scope_name = segments[0]
+        scope_skills = registry.list_by_scope(scope_name)
+        if not scope_skills:
+            send_telegram(f"ℹ️ No skills found in scope '{scope_name}'.")
+            return
+        # Use /command for core skills, /<scope>.<command> for others
+        prefix = "" if scope_name == "core" else f"{scope_name}."
+        parts = [f"Skills in {scope_name}\n"]
+        for skill in scope_skills:
+            for cmd in skill.commands:
+                desc = cmd.description or skill.description
+                parts.append(f"  /{prefix}{cmd.name} -- {desc}")
+        send_telegram("\n".join(parts))
+        return
+
+    scope = segments[0]
+    skill_name = segments[1]
+    subcommand = segments[2] if len(segments) > 2 else skill_name
+
+    skill = registry.get(scope, skill_name)
+    if skill is None:
+        send_telegram(f"❌ Skill '{scope}.{skill_name}' not found. /skill to list available skills.")
+        return
+
+    _dispatch_skill(skill, subcommand, skill_args)
+
+
+def _handle_help_command(command_name: str):
+    """Show help for a specific command: /help <command>."""
+    # Strip leading / if user wrote /help /mission
+    command_name = command_name.lstrip("/").lower()
+
+    registry = _get_registry()
+    skill = registry.find_by_command(command_name)
+
+    if skill is None:
+        send_telegram(f"Unknown command: /{command_name}\nUse /help to see all commands.")
+        return
+
+    # find_by_command maps both names and aliases, so the match is guaranteed
+    cmd = next(
+        c for c in skill.commands
+        if c.name == command_name or command_name in c.aliases
+    )
+
+    parts = [f"/{cmd.name}"]
+    desc = cmd.description or skill.description
+    if desc:
+        parts.append(desc)
+    if cmd.aliases:
+        parts.append(f"Aliases: /{', /'.join(cmd.aliases)}")
+    if cmd.usage:
+        parts.append(f"Usage: {cmd.usage}")
+    else:
+        parts.append("No usage defined.")
+
+    send_telegram("\n".join(parts))
+
+
+def _handle_help():
+    """Send the list of available commands — core + dynamic skills."""
+    registry = _get_registry()
+
+    parts = [
+        "Koan -- Commands\n",
+        "CORE",
+        "⏸️ /pause -- pause (alias: /sleep)",
+        "▶️ /resume -- resume after pause (alias: /work, /awake, /start, /restart)",
+        "⏹️ /stop -- stop Koan after current mission",
+        "/help -- this help (use /help <command> for details)",
+        "/skill -- list available skills",
+    ]
+
+    def _fmt(cmd, skill):
+        desc = cmd.description or skill.description
+        aliases = f" (alias: /{', /'.join(cmd.aliases)})" if cmd.aliases else ""
+        return f"/{cmd.name} -- {desc}{aliases}"
+
+    # Add core skill commands inline (core scope = built-in features)
+    for skill in registry.list_by_scope("core"):
+        for cmd in skill.commands:
+            parts.append(_fmt(cmd, skill))
+    parts.append("")
+
+    # Add non-core skill commands under SKILLS section
+    non_core_skills = [s for s in registry.list_all() if s.scope != "core"]
+    if non_core_skills:
+        parts.append("SKILLS")
+        for skill in non_core_skills:
+            for cmd in skill.commands:
+                parts.append(_fmt(cmd, skill))
+        parts.append("")
+
+    parts.extend([
+        "TIPS",
+        "/help <command> -- show usage for a specific command",
+        'Prefix with "mission:" or use an action verb to create a mission:',
+        "  fix the login bug",
+        "  mission: refactor the auth module",
+        "  [project:koan] fix the login bug",
+        "",
+        "Any other message = free conversation.",
+    ])
+    send_telegram("\n".join(parts))
+
+
+def handle_resume():
+    """Resume from pause or quota exhaustion."""
+    pause_file = KOAN_ROOT / ".koan-pause"
+    pause_reason_file = KOAN_ROOT / ".koan-pause-reason"
+    quota_file = KOAN_ROOT / ".koan-quota-reset"  # Legacy, kept for compat
+
+    if pause_file.exists():
+        # Read pause reason and reset info for better messaging
+        reason = "manual"
+        reset_timestamp = None
+        reset_display = ""
+
+        if pause_reason_file.exists():
+            lines = pause_reason_file.read_text().strip().split("\n")
+            reason = lines[0] if lines else "manual"
+            if len(lines) > 1:
+                try:
+                    reset_timestamp = int(lines[1])
+                except ValueError:
+                    pass
+            if len(lines) > 2:
+                reset_display = lines[2]
+
+        pause_file.unlink(missing_ok=True)
+        pause_reason_file.unlink(missing_ok=True)
+
+        if reason == "quota":
+            # Check if we're resuming before the reset time
+            if reset_timestamp and time.time() < reset_timestamp:
+                from app.reset_parser import time_until_reset
+                remaining = time_until_reset(reset_timestamp)
+                send_telegram(f"▶️ Unpaused (was: quota exhausted). Note: reset is in ~{remaining}. Run loop continues anyway.")
+            else:
+                send_telegram("▶️ Unpaused (was: quota exhausted). Quota should be reset. Run loop continues.")
+        elif reason == "max_runs":
+            send_telegram("▶️ Unpaused (was: max_runs). Run counter reset, loop continues.")
+        else:
+            send_telegram("▶️ Unpaused. Missions resume next cycle.")
+        return
+
+    # Legacy fallback: old .koan-quota-reset file (can be removed in future)
+    if not quota_file.exists():
+        send_telegram("ℹ️ No pause or quota hold detected. /status to check.")
+        return
+
+    try:
+        lines = quota_file.read_text().strip().split("\n")
+        reset_info = lines[0] if lines else "unknown time"
+        paused_at = int(lines[1]) if len(lines) > 1 else 0
+
+        hours_since_pause = (time.time() - paused_at) / 3600
+        likely_reset = hours_since_pause >= 2
+
+        if likely_reset:
+            quota_file.unlink(missing_ok=True)
+            send_telegram(f"▶️ Quota likely reset ({reset_info}, paused {hours_since_pause:.1f}h ago). Restart with: make run")
+        else:
+            send_telegram(f"⏳ Quota not reset yet ({reset_info}). Paused {hours_since_pause:.1f}h ago. Check back later.")
+    except Exception as e:
+        log("error", f"Error checking quota reset: {e}")
+        send_telegram("⚠️ Error checking quota. /status or check manually.")
+
+
+def handle_mission(text: str):
+    """Append to missions.md with optional project tag."""
+    from app.missions import extract_now_flag
+
+    # Check for --now flag in first 5 words (queue at top instead of bottom)
+    urgent, text = extract_now_flag(text)
+
+    # Parse project tag if present
+    project, mission_text = _parse_project(text)
+
+    # Auto-detect project from first word (e.g. "koan do something")
+    if not project:
+        project, detected_text = detect_project_from_text(text)
+        if project:
+            mission_text = detected_text
+
+    # Clean up the mission prefix
+    if mission_text.lower().startswith("mission:"):
+        mission_text = mission_text[8:].strip()
+    elif mission_text.lower().startswith("mission :"):
+        mission_text = mission_text[9:].strip()
+
+    # Format mission entry with project tag if specified
+    if project:
+        mission_entry = f"- [project:{project}] {mission_text}"
+    else:
+        mission_entry = f"- {mission_text}"
+
+    # Append to missions.md under pending section (with file locking)
+    insert_pending_mission(MISSIONS_FILE, mission_entry, urgent=urgent)
+
+    # Acknowledge with project info
+    ack_msg = "✅ Mission received"
+    if urgent:
+        ack_msg += " (priority)"
+    if project:
+        ack_msg += f" (project: {project})"
+    ack_msg += f":\n\n{mission_text[:500]}"
+    send_telegram(ack_msg)
+    log("mission", f"Mission queued: [{project or 'default'}] {mission_text[:60]}")
