+++ koan/tests/test_priority_skill.py
@@ -150,48 +150,48 @@ class TestPriorityHandler:
 class TestPriorityCommandRouting:
     """Test that /priority routes to the priority skill via awake."""
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_priority_routes_via_skill(self, mock_send, tmp_path):
-        from app.awake import handle_command
+        from app.command_handlers import handle_command
 
         missions_file = tmp_path / "missions.md"
         missions_file.write_text(
             "# Missions\n\n## Pending\n\n- first\n- second\n- third\n\n"
             "## In Progress\n\n## Done\n"
         )
-        with patch("app.awake.KOAN_ROOT", tmp_path), \
-             patch("app.awake.INSTANCE_DIR", tmp_path), \
-             patch("app.awake.MISSIONS_FILE", missions_file):
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path), \
+             patch("app.command_handlers.INSTANCE_DIR", tmp_path), \
+             patch("app.command_handlers.MISSIONS_FILE", missions_file):
             handle_command("/priority 3")
         mock_send.assert_called_once()
         output = mock_send.call_args[0][0]
         assert "third" in output
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_priority_bare_shows_queue(self, mock_send, tmp_path):
-        from app.awake import handle_command
+        from app.command_handlers import handle_command
 
         missions_file = tmp_path / "missions.md"
         missions_file.write_text(
             "# Missions\n\n## Pending\n\n- task A\n\n"
             "## In Progress\n\n## Done\n"
         )
-        with patch("app.awake.KOAN_ROOT", tmp_path), \
-             patch("app.awake.INSTANCE_DIR", tmp_path), \
-             patch("app.awake.MISSIONS_FILE", missions_file):
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path), \
+             patch("app.command_handlers.INSTANCE_DIR", tmp_path), \
+             patch("app.command_handlers.MISSIONS_FILE", missions_file):
             handle_command("/priority")
         mock_send.assert_called_once()
         output = mock_send.call_args[0][0]
         assert "task A" in output
         assert "Usage" in output
 
-    @patch("app.awake.send_telegram")
+    @patch("app.command_handlers.send_telegram")
     def test_priority_appears_in_help(self, mock_send, tmp_path):
         """Verify /priority is included in /help output via skill discovery."""
-        from app.awake import handle_command
+        from app.command_handlers import handle_command
 
-        with patch("app.awake.KOAN_ROOT", tmp_path), \
-             patch("app.awake.INSTANCE_DIR", tmp_path):
+        with patch("app.command_handlers.KOAN_ROOT", tmp_path), \
+             patch("app.command_handlers.INSTANCE_DIR", tmp_path):
             handle_command("/help")
         mock_send.assert_called_once()
         help_text = mock_send.call_args[0][0]
